Fix compilation error

+++ src/hotspot/share/gc/shenandoah/shenandoahArguments.cpp
@@ -55,7 +55,7 @@ void ShenandoahArguments::initialize() {
   FLAG_SET_DEFAULT(ShenandoahVerifyOptoBarriers,     false);
 #endif
   if (UseLargePages) {
-    int large_page_size = os::large_page_size();
+    size_t large_page_size = os::large_page_size();
     if ((align_up(MaxHeapSize, large_page_size) / large_page_size) < ShenandoahHeapRegion::MIN_NUM_REGIONS) {
       warning("Large pages size (" SIZE_FORMAT "K) is too large to afford page-sized regions, disabling uncommit",
               os::large_page_size() / K);
